# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AVM CARGO API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 7 */
  phone: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  phone: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  client: 'client',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  name: string;
  phone: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
}

export type PackageStatus = typeof PackageStatus[keyof typeof PackageStatus];


export const PackageStatus = {
  created: 'created',
  accepted_china: 'accepted_china',
  departed_china: 'departed_china',
  arrived_almaty: 'arrived_almaty',
  departed_almaty: 'departed_almaty',
  arrived_city: 'arrived_city',
  ready_pickup: 'ready_pickup',
  delivered: 'delivered',
} as const;

export interface PackageInput {
  /** @minLength 1 */
  trackingNumber: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  deliveryCost?: number | null;
  /** @nullable */
  userId?: number | null;
}

export interface PackageUpdate {
  trackingNumber?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  deliveryCost?: number | null;
  /** @nullable */
  adminComment?: string | null;
  archived?: boolean;
}

export interface StatusInput {
  status: PackageStatus;
}

export interface StatusHistory {
  id: number;
  packageId: number;
  status: PackageStatus;
  changedAt: string;
  /** @nullable */
  changedBy?: number | null;
}

export interface Package {
  id: number;
  trackingNumber: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  deliveryCost?: number | null;
  status: PackageStatus;
  /** @nullable */
  adminComment?: string | null;
  archived: boolean;
  userId: number;
  createdAt: string;
  updatedAt: string;
  user?: User;
}

export interface PackageDetail {
  id: number;
  trackingNumber: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  deliveryCost?: number | null;
  status: PackageStatus;
  /** @nullable */
  adminComment?: string | null;
  archived: boolean;
  userId: number;
  createdAt: string;
  updatedAt: string;
  user?: User;
  history: StatusHistory[];
}

export interface Stats {
  total: number;
  active: number;
  readyPickup: number;
  archived: number;
  recentChanges: Package[];
}

export interface ImportResult {
  updated: number;
  created: number;
  errors: string[];
}

export interface ExportData {
  packages: Package[];
}

export interface Notification {
  id: number;
  userId: number;
  /** @nullable */
  packageId?: number | null;
  message: string;
  read: boolean;
  createdAt: string;
  package?: Package;
}

export type ListPackagesParams = {
/**
 * Filter by archived status
 */
archived?: boolean;
/**
 * Filter by status
 */
status?: string;
/**
 * Search by tracking number, client name, or phone
 */
search?: string;
/**
 * Filter by user ID (admin only)
 * @nullable
 */
userId?: number | null;
};