# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AVM CARGO API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RegisterInput:
      type: object
      required: [name, phone, password]
      properties:
        name:
          type: string
          minLength: 2
        phone:
          type: string
          minLength: 7
        password:
          type: string
          minLength: 6
    LoginInput:
      type: object
      required: [phone, password]
      properties:
        phone:
          type: string
        password:
          type: string
    User:
      type: object
      required: [id, name, phone, role, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [client, admin]
        createdAt:
          type: string
          format: date-time
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"

    # Package schemas
    PackageStatus:
      type: string
      enum:
        - created
        - accepted_china
        - departed_china
        - arrived_almaty
        - departed_almaty
        - arrived_city
        - ready_pickup
        - delivered
...
    Package:
      type: object
      required: [id, trackingNumber, status, archived, userId, createdAt, updatedAt]
      properties:
        id:
          type: number
        trackingNumber:
          type: string
        description:
          type: ["string", "null"]
        weight:
          type: ["number", "null"]
        deliveryCost:
          type: ["number", "null"]
        status:
          $ref: "#/components/schemas/PackageStatus"
        adminComment:
          type: ["string", "null"]
        archived:
          type: boolean
        userId:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
...
    Notification:
      type: object
      required: [id, userId, message, read, createdAt]
      properties:
        id:
          type: number
        userId:
          type: number
        packageId:
          type: ["number", "null"]
        message:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
        package:
          $ref: "#/components/schemas/Package"