# lib/api-zod/src/generated/types/package.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AVM CARGO API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PackageStatus } from './packageStatus';
import type { User } from './user';

export interface Package {
  id: number;
  trackingNumber: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  deliveryCost?: number | null;
  status: PackageStatus;
  /** @nullable */
  adminComment?: string | null;
  archived: boolean;
  userId: number;
  createdAt: Date;
  updatedAt: Date;
  user?: User;
}