# lib/api-zod/src/generated/types/packageStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AVM CARGO API specification
 * OpenAPI spec version: 0.1.0
 */

export type PackageStatus = typeof PackageStatus[keyof typeof PackageStatus];


export const PackageStatus = {
  created: 'created',
  accepted_china: 'accepted_china',
  departed_china: 'departed_china',
  arrived_almaty: 'arrived_almaty',
  departed_almaty: 'departed_almaty',
  arrived_city: 'arrived_city',
  ready_pickup: 'ready_pickup',
  delivered: 'delivered',
} as const;