# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AVM CARGO API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 7 */
  phone: string;
  /** @minLength 6 */
  password: string;
}