# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AVM CARGO API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  name: string;
  phone: string;
  role: UserRole;
  createdAt: Date;
}